refactor: split artifactStore into rendering and retention modules
Extracts pure rendering/formatting functions (509 lines) into
artifactRendering.ts and retention/cleanup/protection logic (958 lines)
into artifactRetention.ts. artifactStore.ts reduced from 2016→598 lines.
Backward-compatible via export * re-exports.
