+++ ralph-codex-vscode-starter/src/ralph/artifactStore.ts
@@ -1,15 +1,21 @@
 import * as fs from 'fs/promises';
 import * as path from 'path';
-import { pathExists, readJsonRecord } from '../util/fs';
 import { stableJson } from './integrity';
 import {
+  renderPreflightSummary,
+  renderIterationSummary,
+  renderIntegrityFailureSummary,
+  renderProvenanceSummary,
+  latestResultFromIteration
+} from './artifactRendering';
+import { cleanupProvenanceBundles } from './artifactRetention';
+import type {
   RalphCliInvocation,
   RalphDiffSummary,
   RalphExecutionPlan,
   RalphIntegrityFailure,
   RalphIterationResult,
   RalphPersistedPreflightReport,
-  RalphPreflightDiagnostic,
   RalphPreflightReport,
   RalphPromptEvidence,
   RalphPromptKind,
@@ -20,6 +26,10 @@ import {
   RalphTaskRemediationArtifact,
   RalphVerificationResult
 } from './types';
+
+// Re-export submodules for backward compatibility.
+export * from './artifactRendering';
+export * from './artifactRetention';
 
 export interface RalphIterationArtifactPaths {
   directory: string;
@@ -70,58 +80,9 @@ export interface RalphLatestArtifactPaths {
   latestProvenanceFailurePath: string;
 }
 
-export interface RalphProvenanceRetentionSummary {
-  deletedBundleIds: string[];
-  retainedBundleIds: string[];
-  protectedBundleIds: string[];
-}
-
-export interface RalphGeneratedArtifactRetentionSummary {
-  deletedIterationDirectories: string[];
-  retainedIterationDirectories: string[];
-  protectedRetainedIterationDirectories: string[];
-  deletedPromptFiles: string[];
-  retainedPromptFiles: string[];
-  protectedRetainedPromptFiles: string[];
-  deletedRunArtifactBaseNames: string[];
-  retainedRunArtifactBaseNames: string[];
-  protectedRetainedRunArtifactBaseNames: string[];
-  deletedHandoffFiles?: string[];
-  retainedHandoffFiles?: string[];
-}
-
 export interface RalphLatestArtifactRepairSummary {
   repairedLatestArtifactPaths: string[];
   staleLatestArtifactPaths: string[];
-}
-
-interface RalphBundleRetentionInspection {
-  bundleIds: string[];
-  retainedBundleIds: string[];
-  protectedBundleIds: string[];
-}
-
-interface RalphGeneratedArtifactRetentionInspection {
-  iterationDirectories: { iteration: number; name: string }[];
-  promptFiles: { iteration: number; name: string }[];
-  runArtifacts: { baseName: string; iteration: number; fileNames: string[] }[];
-  handoffFiles: { iteration: number; name: string }[];
-  protectedArtifacts: RalphProtectedGeneratedArtifacts;
-  iterationDirectoryDecision: {
-    retainedNames: Set<string>;
-    protectedRetainedNames: string[];
-  };
-  promptFileDecision: {
-    retainedNames: Set<string>;
-    protectedRetainedNames: string[];
-  };
-  runArtifactDecision: {
-    retainedNames: Set<string>;
-    protectedRetainedNames: string[];
-  };
-  handoffFileDecision: {
-    retainedNames: Set<string>;
-  };
 }
 
 export type RalphGeneratedArtifactProtectionScope = 'currentAndLatest' | 'fullStateAndLatest';
@@ -209,839 +170,30 @@ export const PROTECTED_GENERATED_LATEST_POINTER_REFERENCES = {
   ]
 } as const;
 
-interface RalphProtectedGeneratedArtifacts {
-  iterationDirectories: Set<string>;
-  promptFiles: Set<string>;
-  runArtifactBaseNames: Set<string>;
-}
-
-interface RalphLatestGeneratedArtifactRecords {
-  latestResult: Record<string, unknown> | null;
-  latestPreflightReport: Record<string, unknown> | null;
-  latestPromptEvidence: Record<string, unknown> | null;
-  latestExecutionPlan: Record<string, unknown> | null;
-  latestCliInvocation: Record<string, unknown> | null;
-  latestProvenanceBundle: Record<string, unknown> | null;
-  latestProvenanceFailure: Record<string, unknown> | null;
-  latestSummaryText: string | null;
-  latestPreflightSummaryText: string | null;
-  latestProvenanceSummaryText: string | null;
-}
-
 export interface RalphPreflightArtifactPaths {
   directory: string;
   reportPath: string;
   summaryPath: string;
 }
 
-function formatOptional(value: string | null | undefined): string {
-  return value && value.trim().length > 0 ? value : 'none';
+export interface RalphProvenanceRetentionSummary {
+  deletedBundleIds: string[];
+  retainedBundleIds: string[];
+  protectedBundleIds: string[];
 }
 
-function bulletList(values: string[]): string {
-  return values.length > 0 ? values.map((value) => `- ${value}`).join('\n') : '- none';
-}
-
-function formatDiagnosticLine(diagnostic: RalphPreflightDiagnostic): string {
-  return `- ${diagnostic.severity}: ${diagnostic.message}`;
-}
-
-function formatTrustLevel(value: RalphProvenanceTrustLevel): string {
-  return value === 'verifiedCliExecution'
-    ? 'verified CLI execution'
-    : 'prepared prompt only';
-}
-
-function sortByIterationDesc<T extends { iteration: number; name: string }>(left: T, right: T): number {
-  return right.iteration - left.iteration || right.name.localeCompare(left.name);
-}
-
-function retainedNamesByNewestAndProtected<T extends { iteration: number }>(
-  entries: readonly T[],
-  retentionCount: number,
-  protectedNames: Iterable<string>,
-  getName: (entry: T) => string
-): Set<string> {
-  // Protection augments the newest-N retention window; it never displaces newer entries.
-  const retained = new Set(entries.slice(0, retentionCount).map((entry) => getName(entry)));
-  for (const name of protectedNames) {
-    retained.add(name);
-  }
-  return retained;
-}
-
-function retentionDecisionByNewestAndProtected<T extends { iteration: number }>(
-  entries: readonly T[],
-  retentionCount: number,
-  protectedNames: Iterable<string>,
-  getName: (entry: T) => string
-): {
-  retainedNames: Set<string>;
-  protectedRetainedNames: string[];
-} {
-  const newestWindowNames = new Set(entries.slice(0, retentionCount).map((entry) => getName(entry)));
-  const retainedNames = retainedNamesByNewestAndProtected(entries, retentionCount, protectedNames, getName);
-  const protectedRetainedNames = entries
-    .filter((entry) => {
-      const name = getName(entry);
-      return retainedNames.has(name) && !newestWindowNames.has(name);
-    })
-    .map((entry) => getName(entry));
-
-  return {
-    retainedNames,
-    protectedRetainedNames
-  };
-}
-
-function parseIterationDirectoryName(name: string): { iteration: number; name: string } | null {
-  const match = /^iteration-(\d+)$/.exec(name);
-  if (!match) {
-    return null;
-  }
-
-  return {
-    name,
-    iteration: Number.parseInt(match[1], 10)
-  };
-}
-
-function parsePromptFileName(name: string): { iteration: number; name: string } | null {
-  const match = /^.+-(\d+)\.prompt\.md$/.exec(name);
-  if (!match) {
-    return null;
-  }
-
-  return {
-    name,
-    iteration: Number.parseInt(match[1], 10)
-  };
-}
-
-function parseRunArtifactFileName(name: string): { baseName: string; iteration: number; name: string } | null {
-  const match = /^(.+)-(\d+)\.(transcript|last-message)\.md$/.exec(name);
-  if (!match) {
-    return null;
-  }
-
-  return {
-    name,
-    baseName: `${match[1]}-${match[2]}`,
-    iteration: Number.parseInt(match[2], 10)
-  };
-}
-
-function createProtectedGeneratedArtifacts(): RalphProtectedGeneratedArtifacts {
-  return {
-    iterationDirectories: new Set<string>(),
-    promptFiles: new Set<string>(),
-    runArtifactBaseNames: new Set<string>()
-  };
-}
-
-function isPathWithin(rootDir: string, targetPath: string): boolean {
-  if (!path.isAbsolute(targetPath)) {
-    return false;
-  }
-
-  const relative = path.relative(rootDir, targetPath);
-  return !relative.startsWith('..') && !path.isAbsolute(relative);
-}
-
-function readPathReference(record: Record<string, unknown> | null, key: string): string | null {
-  const value = record?.[key];
-  return typeof value === 'string' && value.trim().length > 0 ? value.trim() : null;
-}
-
-function readRecordReference(record: Record<string, unknown> | null, key: string): Record<string, unknown> | null {
-  const value = record?.[key];
-  return typeof value === 'object' && value !== null && !Array.isArray(value)
-    ? value as Record<string, unknown>
-    : null;
-}
-
-function compactPathReferences(references: Array<string | null>): string[] {
-  return references.filter((value): value is string => Boolean(value));
-}
-
-function parsePromptEvidenceIdentity(record: Record<string, unknown> | null): {
-  kind: RalphPromptKind;
-  iteration: number;
-} | null {
-  if (!record) {
-    return null;
-  }
-
-  const kind = record.kind;
-  const iteration = record.iteration;
-  if ((kind !== 'bootstrap'
-    && kind !== 'iteration'
-    && kind !== 'replenish-backlog'
-    && kind !== 'fix-failure'
-    && kind !== 'continue-progress'
-    && kind !== 'human-review-handoff')
-    || typeof iteration !== 'number'
-    || !Number.isFinite(iteration)
-    || iteration < 1) {
-    return null;
-  }
-
-  return {
-    kind,
-    iteration: Math.floor(iteration)
-  };
-}
-
-function derivedLatestPromptEvidenceReferences(
-  record: Record<string, unknown> | null,
-  dirs: {
-    artifactRootDir: string;
-    promptDir: string;
-  }
-): string[] {
-  const identity = parsePromptEvidenceIdentity(record);
-  if (!identity) {
-    return [];
-  }
-
-  const paddedIteration = String(identity.iteration).padStart(3, '0');
-  return [
-    path.join(dirs.artifactRootDir, `iteration-${paddedIteration}`),
-    path.join(dirs.promptDir, `${identity.kind}-${paddedIteration}.prompt.md`)
-  ];
-}
-
-function readRecordArray(record: Record<string, unknown> | null, key: string): Record<string, unknown>[] {
-  const value = record?.[key];
-  if (!Array.isArray(value)) {
-    return [];
-  }
-
-  return value.filter((candidate): candidate is Record<string, unknown> =>
-    typeof candidate === 'object' && candidate !== null && !Array.isArray(candidate)
-  );
-}
-
-function addProtectedGeneratedArtifactPath(
-  protectedArtifacts: RalphProtectedGeneratedArtifacts,
-  input: {
-    artifactRootDir: string;
-    promptDir: string;
-    runDir: string;
-    targetPath: string;
-  }
-): void {
-  const normalizedPath = path.normalize(input.targetPath);
-
-  if (isPathWithin(input.artifactRootDir, normalizedPath)) {
-    const relative = path.relative(input.artifactRootDir, normalizedPath);
-    const [firstSegment] = relative.split(path.sep);
-    if (firstSegment && parseIterationDirectoryName(firstSegment)) {
-      protectedArtifacts.iterationDirectories.add(firstSegment);
-    }
-  }
-
-  const baseName = path.basename(normalizedPath);
-
-  if (isPathWithin(input.promptDir, normalizedPath)) {
-    const parsedPrompt = parsePromptFileName(baseName);
-    if (parsedPrompt) {
-      protectedArtifacts.promptFiles.add(parsedPrompt.name);
-    }
-  }
-
-  if (isPathWithin(input.runDir, normalizedPath)) {
-    const parsedRunArtifact = parseRunArtifactFileName(baseName);
-    if (parsedRunArtifact) {
-      protectedArtifacts.runArtifactBaseNames.add(parsedRunArtifact.baseName);
-    }
-  }
-}
-
-function addProtectedGeneratedArtifactPaths(
-  protectedArtifacts: RalphProtectedGeneratedArtifacts,
-  input: {
-    artifactRootDir: string;
-    promptDir: string;
-    runDir: string;
-  },
-  targetPaths: Iterable<string>
-): void {
-  for (const targetPath of targetPaths) {
-    addProtectedGeneratedArtifactPath(protectedArtifacts, {
-      ...input,
-      targetPath
-    });
-  }
-}
-
-function readStateRunReference(candidate: Record<string, unknown> | null): {
-  iteration: number | null;
-  promptPath: string | null;
-  transcriptPath: string | null;
-  lastMessagePath: string | null;
-} | null {
-  if (!candidate) {
-    return null;
-  }
-
-  const iteration = typeof candidate.iteration === 'number'
-    && Number.isFinite(candidate.iteration)
-    && candidate.iteration >= 1
-    ? Math.floor(candidate.iteration)
-    : null;
-  const promptPath = readPathReference(candidate, 'promptPath');
-  const transcriptPath = readPathReference(candidate, 'transcriptPath');
-  const lastMessagePath = readPathReference(candidate, 'lastMessagePath');
-
-  if (!promptPath && !transcriptPath && !lastMessagePath && iteration === null) {
-    return null;
-  }
-
-  return {
-    iteration,
-    promptPath,
-    transcriptPath,
-    lastMessagePath
-  };
-}
-
-function derivedIterationDirectoryPath(artifactRootDir: string, iteration: number): string {
-  return path.join(artifactRootDir, `iteration-${String(iteration).padStart(3, '0')}`);
-}
-
-function currentStateArtifactReferences(input: {
-  stateRecord: Record<string, unknown> | null;
-  artifactRootDir: string;
-  includeHistory?: boolean;
-}): string[] {
-  if (!input.stateRecord) {
-    return [];
-  }
-
-  const runHistory = readRecordArray(input.stateRecord, 'runHistory')
-    .map((record) => readStateRunReference(record))
-    .filter((record): record is NonNullable<ReturnType<typeof readStateRunReference>> => record !== null);
-  const lastRun = readStateRunReference(readRecordReference(input.stateRecord, 'lastRun')) ?? runHistory.at(-1) ?? null;
-  const effectiveRunHistory = lastRun ? [...runHistory, lastRun] : runHistory;
-  const iterationHistory = readRecordArray(input.stateRecord, 'iterationHistory');
-  const effectiveIterationHistory = iterationHistory.length > 0
-    ? iterationHistory
-    : effectiveRunHistory.flatMap((runRecord) => {
-      if (runRecord.iteration === null) {
-        return [];
-      }
-
-      return [{
-        artifactDir: derivedIterationDirectoryPath(input.artifactRootDir, runRecord.iteration),
-        promptPath: runRecord.promptPath,
-        execution: {
-          transcriptPath: runRecord.transcriptPath,
-          lastMessagePath: runRecord.lastMessagePath
-        }
-      }];
-    });
-  const lastIteration = readRecordReference(input.stateRecord, 'lastIteration') ?? effectiveIterationHistory.at(-1) ?? null;
-  const lastIterationExecution = readRecordReference(lastIteration, 'execution');
-  const references = compactPathReferences([
-    readPathReference(input.stateRecord, 'lastPromptPath'),
-    readPathReference(lastRun, 'promptPath'),
-    readPathReference(lastRun, 'transcriptPath'),
-    readPathReference(lastRun, 'lastMessagePath'),
-    readPathReference(lastIteration, 'artifactDir'),
-    readPathReference(lastIteration, 'promptPath'),
-    readPathReference(lastIterationExecution, 'transcriptPath'),
-    readPathReference(lastIterationExecution, 'lastMessagePath')
-  ]);
-
-  if (input.includeHistory ?? true) {
-    for (const runRecord of runHistory) {
-      references.push(...compactPathReferences([
-        readPathReference(runRecord, 'promptPath'),
-        readPathReference(runRecord, 'transcriptPath'),
-        readPathReference(runRecord, 'lastMessagePath')
-      ]));
-    }
-
-    for (const iterationRecord of effectiveIterationHistory) {
-      const executionRecord = readRecordReference(iterationRecord, 'execution');
-      references.push(...compactPathReferences([
-        readPathReference(iterationRecord, 'artifactDir'),
-        readPathReference(iterationRecord, 'promptPath'),
-        readPathReference(executionRecord, 'transcriptPath'),
-        readPathReference(executionRecord, 'lastMessagePath')
-      ]));
-    }
-  }
-
-  return references;
-}
-
-function latestArtifactReferences(
-  records: RalphLatestGeneratedArtifactRecords,
-  dirs: {
-    artifactRootDir: string;
-    promptDir: string;
-  }
-): string[] {
-  const latestResultReferenceFields = PROTECTED_GENERATED_LATEST_POINTER_REFERENCES['latest-result.json'];
-  const latestPreflightReferenceFields = PROTECTED_GENERATED_LATEST_POINTER_REFERENCES['latest-preflight-report.json'];
-  const latestExecutionPlanReferenceFields = PROTECTED_GENERATED_LATEST_POINTER_REFERENCES['latest-execution-plan.json'];
-  const latestCliInvocationReferenceFields = PROTECTED_GENERATED_LATEST_POINTER_REFERENCES['latest-cli-invocation.json'];
-  const latestProvenanceBundleReferenceFields = PROTECTED_GENERATED_LATEST_POINTER_REFERENCES['latest-provenance-bundle.json'];
-  const latestProvenanceFailureReferenceFields = PROTECTED_GENERATED_LATEST_POINTER_REFERENCES['latest-provenance-failure.json'];
-
-  return compactPathReferences([
-    ...latestResultReferenceFields.map((field) => readPathReference(records.latestResult, field)),
-    ...latestPreflightReferenceFields.map((field) => readPathReference(records.latestPreflightReport, field)),
-    ...latestExecutionPlanReferenceFields.map((field) => readPathReference(records.latestExecutionPlan, field)),
-    ...latestCliInvocationReferenceFields.map((field) => readPathReference(records.latestCliInvocation, field)),
-    ...latestProvenanceBundleReferenceFields.map((field) => readPathReference(records.latestProvenanceBundle, field)),
-    ...latestProvenanceFailureReferenceFields.map((field) => readPathReference(records.latestProvenanceFailure, field)),
-    ...derivedLatestPromptEvidenceReferences(records.latestPromptEvidence, dirs),
-    ...derivedLatestSurfaceReferences({
-      artifactRootDir: dirs.artifactRootDir,
-      latestSummaryText: records.latestSummaryText,
-      latestPreflightSummaryText: records.latestPreflightSummaryText,
-      latestProvenanceSummaryText: records.latestProvenanceSummaryText
-    })
-  ]);
-}
-
-function artifactReferenceLines(paths: RalphIterationArtifactPaths, diffSummary: RalphDiffSummary | null): string[] {
-  const lines = [
-    `- Prompt: ${paths.promptPath}`,
-    `- Prompt evidence: ${paths.promptEvidencePath}`,
-    `- Execution plan: ${paths.executionPlanPath}`,
-    `- Completion report: ${paths.completionReportPath}`,
-    `- Execution summary: ${paths.executionSummaryPath}`,
-    `- Verifier summary: ${paths.verifierSummaryPath}`,
-    `- Iteration result: ${paths.iterationResultPath}`,
-    `- Remediation proposal: ${paths.remediationPath}`,
-    `- Stdout: ${paths.stdoutPath}`,
-    `- Stderr: ${paths.stderrPath}`,
-    `- CLI invocation: ${paths.cliInvocationPath}`
-  ];
-
-  if (diffSummary) {
-    lines.push(`- Diff summary: ${paths.diffSummaryPath}`);
-  }
-  if (diffSummary?.beforeStatusPath) {
-    lines.push(`- Git status before: ${paths.gitStatusBeforePath}`);
-  }
-  if (diffSummary?.afterStatusPath) {
-    lines.push(`- Git status after: ${paths.gitStatusAfterPath}`);
-  }
-
-  return lines;
-}
-
-function renderPreflightSummary(report: RalphPersistedPreflightReport): string {
-  const headline = report.ready
-    ? 'Preflight completed without blocking errors.'
-    : 'Preflight blocked before Codex execution started.';
-  const diagnostics = report.diagnostics.length > 0
-    ? report.diagnostics.map(formatDiagnosticLine)
-    : ['- ok'];
-
-  const handoffLine = report.sessionHandoff
-    ? `- Session handoff: ${report.sessionHandoff.agentId}-${String(report.sessionHandoff.iteration).padStart(3, '0')}.json | ${report.sessionHandoff.humanSummary}`
-    : '- Session handoff: none';
-
-  return [
-    `# Ralph Preflight ${report.iteration}`,
-    '',
-    `- Agent ID: ${formatOptional(report.agentId)}`,
-    `- Provenance ID: ${report.provenanceId}`,
-    `- Trust level: ${formatTrustLevel(report.trustLevel)}`,
-    `- Ready: ${report.ready ? 'yes' : 'no'}`,
-    `- Prompt kind: ${report.promptKind}`,
-    `- Prompt target: ${report.promptTarget}`,
-    `- Selected task: ${formatOptional(report.selectedTaskId)}${report.selectedTaskTitle ? ` - ${report.selectedTaskTitle}` : ''}`,
-    `- Task validation hint: ${formatOptional(report.taskValidationHint)}`,
-    `- Effective validation command: ${formatOptional(report.effectiveValidationCommand)}`,
-    `- Validation normalized from: ${formatOptional(report.normalizedValidationCommandFrom)}`,
-    `- Validation: ${formatOptional(report.validationCommand)}`,
-    `- Summary: ${report.summary}`,
-    `- Active claim state: ${report.activeClaimSummary ?? 'none'}`,
-    handoffLine,
-    `- Report: ${report.reportPath}`,
-    '',
-    headline,
-    '',
-    '## Diagnostics',
-    ...diagnostics
-  ].join('\n');
-}
-
-function renderIterationSummary(input: {
-  result: RalphIterationResult;
-  paths: RalphIterationArtifactPaths;
-  verifiers: RalphVerificationResult[];
-  diffSummary: RalphDiffSummary | null;
-}): string {
-  const { result, paths, diffSummary, verifiers } = input;
-  const verifierLines = verifiers.map((verifier) => {
-    const location = verifier.artifactPath ? ` (${verifier.artifactPath})` : '';
-    return `${verifier.verifier}: ${verifier.status} - ${verifier.summary}${location}`;
-  });
-  const diffLines = diffSummary
-    ? [
-      `- Summary: ${diffSummary.summary}`,
-      `- Git available: ${diffSummary.gitAvailable ? 'yes' : 'no'}`,
-      `- Changed files: ${diffSummary.changedFileCount}`,
-      `- Relevant changed files: ${diffSummary.relevantChangedFileCount}`,
-      `- Suggested checkpoint ref: ${diffSummary.suggestedCheckpointRef ?? 'none'}`
-    ]
-    : ['- none'];
-
-  return [
-    `# Ralph Iteration ${result.iteration}`,
-    '',
-    '## Outcome',
-    `- Agent ID: ${formatOptional(result.agentId)}`,
-    `- Provenance ID: ${formatOptional(result.provenanceId)}`,
-    `- Selected task: ${formatOptional(result.selectedTaskId)}${result.selectedTaskTitle ? ` - ${result.selectedTaskTitle}` : ''}`,
-    `- Prompt kind: ${result.promptKind}`,
-    `- Target mode: ${result.executionIntegrity?.promptTarget ?? 'unknown'}`,
-    `- Template: ${result.executionIntegrity?.templatePath ?? 'unknown'}`,
-    `- Reasoning effort: ${result.executionIntegrity?.reasoningEffort ?? 'unknown'}`,
-    `- Execution: ${result.executionStatus}`,
-    `- Execution message: ${result.execution.message ?? 'none'}`,
-    `- Verification: ${result.verificationStatus}`,
-    `- Classification: ${result.completionClassification} (selected task)`,
-    `- Backlog remaining: ${result.backlog.remainingTaskCount}`,
-    `- Next actionable task available: ${result.backlog.actionableTaskAvailable ? 'yes' : 'no'}`,
-    `- Follow-up action: ${result.followUpAction}`,
-    `- Stop reason: ${formatOptional(result.stopReason)}`,
-    `- Remediation: ${result.remediation?.summary ?? 'none'}`,
-    `- Summary: ${result.summary}`,
-    '',
-    '## Execution Integrity',
-    `- Plan: ${result.executionIntegrity?.executionPlanPath ?? 'none'}`,
-    `- Plan hash: ${result.executionIntegrity?.executionPlanHash ?? 'none'}`,
-    `- Prompt artifact: ${result.executionIntegrity?.promptArtifactPath ?? 'none'}`,
-    `- Prompt hash: ${result.executionIntegrity?.promptHash ?? 'none'}`,
-    `- Workspace root: ${result.executionIntegrity?.rootPolicy?.workspaceRootPath ?? 'none'}`,
-    `- Execution root: ${result.executionIntegrity?.rootPolicy?.executionRootPath ?? 'none'}`,
-    `- Verifier root: ${result.executionIntegrity?.rootPolicy?.verificationRootPath ?? 'none'}`,
-    `- Payload matched rendered artifact: ${result.executionIntegrity?.executionPayloadMatched == null
-      ? 'not executed'
-      : result.executionIntegrity.executionPayloadMatched ? 'yes' : 'no'}`,
-    `- CLI invocation: ${result.executionIntegrity?.cliInvocationPath ?? 'none'}`,
-    `- Integrity issue: ${result.executionIntegrity?.mismatchReason ?? 'none'}`,
-    '',
-    '## Validation',
-    `- Task validation hint: ${formatOptional(result.verification.taskValidationHint)}`,
-    `- Effective validation command: ${formatOptional(result.verification.effectiveValidationCommand)}`,
-    `- Validation command normalized from: ${formatOptional(result.verification.normalizedValidationCommandFrom)}`,
-    `- Primary command: ${formatOptional(result.verification.primaryCommand)}`,
-    `- Failure signature: ${formatOptional(result.verification.validationFailureSignature)}`,
-    verifierLines.length > 0 ? bulletList(verifierLines) : '- none',
-    '',
-    '## Diff',
-    ...diffLines,
-    '',
-    '## Artifact Paths',
-    ...artifactReferenceLines(paths, diffSummary),
-    '',
-    '## Signals',
-    `- No-progress signals: ${result.noProgressSignals.join(', ') || 'none'}`,
-    `- Remediation action: ${result.remediation?.action ?? 'none'}`,
-    `- Remediation evidence: ${result.remediation?.evidence.join(' | ') || 'none'}`,
-    `- Remediation proposal artifact: ${result.remediation ? paths.remediationPath : 'none'}`,
-    `- Completion report status: ${result.completionReportStatus ?? 'none'}`,
-    `- Reconciliation warnings: ${result.reconciliationWarnings?.join(' | ') || 'none'}`,
-    `- Warnings: ${result.warnings.join(' | ') || 'none'}`,
-    `- Errors: ${result.errors.join(' | ') || 'none'}`
-  ].join('\n');
-}
-
-function renderIntegrityFailureSummary(failure: RalphIntegrityFailure): string {
-  return [
-    `# Ralph Provenance Failure ${failure.iteration}`,
-    '',
-    `- Provenance ID: ${failure.provenanceId}`,
-    `- Stage: ${failure.stage}`,
-    `- Prompt kind: ${failure.promptKind}`,
-    `- Prompt target: ${failure.promptTarget}`,
-    `- Trust level: ${formatTrustLevel(failure.trustLevel)}`,
-    `- Summary: ${failure.summary}`,
-    `- Message: ${failure.message}`,
-    '',
-    '## Expected vs Actual',
-    `- Expected execution plan hash: ${failure.expectedExecutionPlanHash ?? 'none'}`,
-    `- Actual execution plan hash: ${failure.actualExecutionPlanHash ?? 'none'}`,
-    `- Expected prompt hash: ${failure.expectedPromptHash ?? 'none'}`,
-    `- Actual prompt hash: ${failure.actualPromptHash ?? 'none'}`,
-    `- Expected payload hash: ${failure.expectedPayloadHash ?? 'none'}`,
-    `- Actual payload hash: ${failure.actualPayloadHash ?? 'none'}`,
-    '',
-    '## Artifact Paths',
-    `- Iteration artifact dir: ${failure.artifactDir}`,
-    `- Execution plan: ${failure.executionPlanPath ?? 'none'}`,
-    `- Prompt artifact: ${failure.promptArtifactPath ?? 'none'}`,
-    `- CLI invocation: ${failure.cliInvocationPath ?? 'none'}`
-  ].join('\n');
-}
-
-function renderProvenanceSummary(bundle: RalphProvenanceBundle): string {
-  const reconciliationWarnings = bundle.reconciliationWarnings ?? [];
-  const epistemicGap = bundle.epistemicGap ?? {
-    trustBoundary: 'The provenance chain stops at the codex exec boundary; model-internal reasoning is not directly observable.',
-    bundleProves: 'The persisted Ralph artifacts and any verifier-observed run outputs recorded in this bundle.',
-    bundleDoesNotProve: 'That the model reasoned correctly internally or that an unverified completion report is true.',
-    modelClaimsPath: bundle.completionReportPath ?? null,
-    modelClaimsStatus: bundle.completionReportStatus ?? null,
-    modelClaimsAreUnverified: (bundle.completionReportPath ?? null) !== null,
-    verifierEvidencePaths: [bundle.executionSummaryPath, bundle.verifierSummaryPath, bundle.iterationResultPath]
-      .filter((item): item is string => typeof item === 'string' && item.length > 0),
-    verifierEvidenceIsAuthoritative: true,
-    reconciliationWarnings,
-    noWarningsMeans: 'No reconciliation warnings means the model claim matched the observable verifier signals, not that the model reasoning was correct.'
-  };
-  const verifierEvidencePaths = epistemicGap.verifierEvidencePaths;
-
-  return [
-    `# Ralph Provenance ${bundle.provenanceId}`,
-    '',
-    `- Agent ID: ${formatOptional(bundle.agentId)}`,
-    `- Iteration: ${bundle.iteration}`,
-    `- Status: ${bundle.status}`,
-    `- Trust level: ${formatTrustLevel(bundle.trustLevel)}`,
-    `- Prompt kind: ${bundle.promptKind}`,
-    `- Prompt target: ${bundle.promptTarget}`,
-    `- Workspace root: ${bundle.rootPolicy.workspaceRootPath}`,
-    `- Inspection root: ${bundle.rootPolicy.inspectionRootPath}`,
-    `- Execution root: ${bundle.rootPolicy.executionRootPath}`,
-    `- Verifier root: ${bundle.rootPolicy.verificationRootPath}`,
-    `- Selected task: ${formatOptional(bundle.selectedTaskId)}${bundle.selectedTaskTitle ? ` - ${bundle.selectedTaskTitle}` : ''}`,
-    `- Summary: ${bundle.summary}`,
-    '',
-    '## Integrity',
-    `- Execution plan hash: ${bundle.executionPlanHash ?? 'none'}`,
-    `- Prompt hash: ${bundle.promptHash ?? 'none'}`,
-    `- Payload matched rendered artifact: ${bundle.executionPayloadMatched == null
-      ? 'not executed'
-      : bundle.executionPayloadMatched ? 'yes' : 'no'}`,
-    `- Integrity issue: ${bundle.mismatchReason ?? 'none'}`,
-    '',
-    '## Bundle Files',
-    `- Bundle manifest: ${bundle.bundleDir ? path.join(bundle.bundleDir, 'provenance-bundle.json') : 'none'}`,
-    `- Preflight report: ${bundle.preflightReportPath}`,
-    `- Preflight summary: ${bundle.preflightSummaryPath}`,
-    `- Prompt artifact: ${bundle.promptArtifactPath ?? 'none'}`,
-    `- Prompt evidence: ${bundle.promptEvidencePath ?? 'none'}`,
-    `- Execution plan: ${bundle.executionPlanPath ?? 'none'}`,
-    `- CLI invocation: ${bundle.cliInvocationPath ?? 'none'}`,
-    `- Iteration result: ${bundle.iterationResultPath ?? 'none'}`,
-    `- Execution summary: ${bundle.executionSummaryPath ?? 'none'}`,
-    `- Verifier summary: ${bundle.verifierSummaryPath ?? 'none'}`,
-    `- Completion report: ${bundle.completionReportPath ?? 'none'}`,
-    `- Provenance failure: ${bundle.provenanceFailurePath ?? 'none'}`,
-    '',
-    '## Model Claims',
-    `- Completion report path: ${epistemicGap.modelClaimsPath ?? 'none'}`,
-    `- Model self-report status: ${epistemicGap.modelClaimsStatus ?? 'none'}`,
-    `- Unverified model claim: ${epistemicGap.modelClaimsAreUnverified ? 'yes' : 'no'}`,
-    `- Reconciliation warnings: ${reconciliationWarnings.join(' | ') || 'none'}`,
-    '',
-    '## Verifier Evidence',
-    ...(verifierEvidencePaths.length > 0
-      ? verifierEvidencePaths.map((evidencePath) => `- ${evidencePath}`)
-      : ['- none']),
-    `- Verifier evidence is authoritative: ${epistemicGap.verifierEvidenceIsAuthoritative ? 'yes' : 'no'}`,
-    '',
-    '## Epistemic Gap',
-    `- Trust boundary: ${epistemicGap.trustBoundary}`,
-    `- This bundle proves: ${epistemicGap.bundleProves}`,
-    `- This bundle does not prove: ${epistemicGap.bundleDoesNotProve}`,
-    `- No-warning interpretation: ${epistemicGap.noWarningsMeans}`,
-    '',
-    '## Canonical Iteration Artifacts',
-    `- Iteration artifact dir: ${bundle.artifactDir}`
-  ].join('\n');
-}
-
-function parseHandoffFileName(name: string): { iteration: number; name: string } | null {
-  const match = /^.+-(\d+)\.json$/.exec(name);
-  if (!match) {
-    return null;
-  }
-
-  return {
-    name,
-    iteration: Number.parseInt(match[1], 10)
-  };
-}
-
-function renderLatestResultSummary(record: Record<string, unknown>): string | null {
-  if (typeof record.iteration !== 'number'
-    || typeof record.promptKind !== 'string'
-    || typeof record.executionStatus !== 'string'
-    || typeof record.verificationStatus !== 'string'
-    || typeof record.completionClassification !== 'string'
-    || typeof record.followUpAction !== 'string'
-    || typeof record.summary !== 'string') {
-    return null;
-  }
-
-  const selectedTaskId = typeof record.selectedTaskId === 'string' ? record.selectedTaskId : null;
-  const selectedTaskTitle = typeof record.selectedTaskTitle === 'string' ? record.selectedTaskTitle : null;
-  const agentId = typeof record.agentId === 'string' ? record.agentId : 'none';
-  const promptTarget = typeof record.promptTarget === 'string' ? record.promptTarget : 'unknown';
-  const templatePath = typeof record.templatePath === 'string' ? record.templatePath : 'unknown';
-  const executionMessage = typeof record.executionMessage === 'string' ? record.executionMessage : 'none';
-  const stopReason = typeof record.stopReason === 'string' ? record.stopReason : 'none';
-  const taskValidationHint = typeof record.taskValidationHint === 'string' ? record.taskValidationHint : 'none';
-  const effectiveValidationCommand = typeof record.effectiveValidationCommand === 'string'
-    ? record.effectiveValidationCommand
-    : 'none';
-  const normalizedValidationCommandFrom = typeof record.normalizedValidationCommandFrom === 'string'
-    ? record.normalizedValidationCommandFrom
-    : 'none';
-  const promptArtifactPath = typeof record.promptArtifactPath === 'string' ? record.promptArtifactPath : 'none';
-  const executionPlanPath = typeof record.executionPlanPath === 'string' ? record.executionPlanPath : 'none';
-  const cliInvocationPath = typeof record.cliInvocationPath === 'string' ? record.cliInvocationPath : 'none';
-  const remediationSummary = typeof (record.remediation as Record<string, unknown> | null)?.summary === 'string'
-    ? (record.remediation as Record<string, unknown>).summary as string
-    : 'none';
-  const summaryPath = typeof record.summaryPath === 'string' ? record.summaryPath : 'none';
-  const transcriptPath = typeof record.transcriptPath === 'string' ? record.transcriptPath : 'none';
-  const lastMessagePath = typeof record.lastMessagePath === 'string' ? record.lastMessagePath : 'none';
-  const executionSummaryPath = typeof record.executionSummaryPath === 'string' ? record.executionSummaryPath : 'none';
-  const verifierSummaryPath = typeof record.verifierSummaryPath === 'string' ? record.verifierSummaryPath : 'none';
-  const iterationResultPath = typeof record.iterationResultPath === 'string' ? record.iterationResultPath : 'none';
-  const stdoutPath = typeof record.stdoutPath === 'string' ? record.stdoutPath : 'none';
-  const stderrPath = typeof record.stderrPath === 'string' ? record.stderrPath : 'none';
-  const diffSummaryPath = typeof record.diffSummaryPath === 'string' ? record.diffSummaryPath : 'none';
-  const remediationPath = typeof record.remediationPath === 'string' ? record.remediationPath : 'none';
-  const completionReportStatus = typeof record.completionReportStatus === 'string' ? record.completionReportStatus : 'none';
-  const promptHash = typeof record.promptHash === 'string' ? record.promptHash : 'none';
-  const executionPlanHash = typeof record.executionPlanHash === 'string' ? record.executionPlanHash : 'none';
-  const artifactDir = typeof record.artifactDir === 'string' ? record.artifactDir : 'none';
-  const backlog = typeof record.backlog === 'object' && record.backlog !== null
-    ? record.backlog as Record<string, unknown>
-    : null;
-  const remainingTaskCount = typeof backlog?.remainingTaskCount === 'number' ? Math.max(0, Math.floor(backlog.remainingTaskCount)) : 0;
-  const actionableTaskAvailable = Boolean(backlog?.actionableTaskAvailable);
-  const warnings = Array.isArray(record.warnings) ? record.warnings.filter((item): item is string => typeof item === 'string') : [];
-  const errors = Array.isArray(record.errors) ? record.errors.filter((item): item is string => typeof item === 'string') : [];
-  const reconciliationWarnings = Array.isArray(record.reconciliationWarnings)
-    ? record.reconciliationWarnings.filter((item): item is string => typeof item === 'string')
-    : [];
-
-  return [
-    `# Ralph Iteration ${Math.floor(record.iteration)}`,
-    '',
-    '## Outcome',
-    `- Agent ID: ${agentId}`,
-    `- Provenance ID: ${typeof record.provenanceId === 'string' ? record.provenanceId : 'none'}`,
-    `- Selected task: ${selectedTaskId ?? 'none'}${selectedTaskTitle ? ` - ${selectedTaskTitle}` : ''}`,
-    `- Prompt kind: ${record.promptKind}`,
-    `- Target mode: ${promptTarget}`,
-    `- Template: ${templatePath}`,
-    `- Execution: ${record.executionStatus}`,
-    `- Execution message: ${executionMessage}`,
-    `- Verification: ${record.verificationStatus}`,
-    `- Classification: ${record.completionClassification} (selected task)`,
-    `- Backlog remaining: ${remainingTaskCount}`,
-    `- Next actionable task available: ${actionableTaskAvailable ? 'yes' : 'no'}`,
-    `- Follow-up action: ${record.followUpAction}`,
-    `- Stop reason: ${stopReason}`,
-    `- Remediation: ${remediationSummary}`,
-    `- Summary: ${record.summary}`,
-    '',
-    '## Execution Integrity',
-    `- Plan: ${executionPlanPath}`,
-    `- Plan hash: ${executionPlanHash}`,
-    `- Prompt artifact: ${promptArtifactPath}`,
-    `- Prompt hash: ${promptHash}`,
-    `- Payload matched rendered artifact: ${record.executionPayloadMatched == null
-      ? 'not recorded'
-      : record.executionPayloadMatched ? 'yes' : 'no'}`,
-    `- CLI invocation: ${cliInvocationPath}`,
-    '',
-    '## Validation',
-    `- Task validation hint: ${taskValidationHint}`,
-    `- Effective validation command: ${effectiveValidationCommand}`,
-    `- Validation command normalized from: ${normalizedValidationCommandFrom}`,
-    '',
-    '## Artifact Paths',
-    `- Prompt: ${typeof record.promptPath === 'string' ? record.promptPath : 'none'}`,
-    `- Prompt evidence: ${typeof record.promptEvidencePath === 'string' ? record.promptEvidencePath : 'none'}`,
-    `- Execution plan: ${executionPlanPath}`,
-    `- Execution summary: ${executionSummaryPath}`,
-    `- Verifier summary: ${verifierSummaryPath}`,
-    `- Iteration result: ${iterationResultPath}`,
-    `- Remediation proposal: ${remediationPath}`,
-    `- Summary: ${summaryPath}`,
-    `- Transcript: ${transcriptPath}`,
-    `- Last message: ${lastMessagePath}`,
-    `- Stdout: ${stdoutPath}`,
-    `- Stderr: ${stderrPath}`,
-    `- Diff summary: ${diffSummaryPath}`,
-    `- Iteration artifact dir: ${artifactDir}`,
-    '',
-    '## Signals',
-    `- Completion report status: ${completionReportStatus}`,
-    `- Reconciliation warnings: ${reconciliationWarnings.join(' | ') || 'none'}`,
-    `- Warnings: ${warnings.join(' | ') || 'none'}`,
-    `- Errors: ${errors.join(' | ') || 'none'}`
-  ].join('\n');
-}
-
-function latestResultFromIteration(input: {
-  result: RalphIterationResult;
-  paths: RalphIterationArtifactPaths;
-  diffSummary: RalphDiffSummary | null;
-}): Record<string, unknown> {
-  return {
-    agentId: input.result.agentId ?? null,
-    provenanceId: input.result.provenanceId ?? null,
-    iteration: input.result.iteration,
-    selectedTaskId: input.result.selectedTaskId,
-    selectedTaskTitle: input.result.selectedTaskTitle,
-    promptKind: input.result.promptKind,
-    promptTarget: input.result.executionIntegrity?.promptTarget ?? null,
-    rootPolicy: input.result.executionIntegrity?.rootPolicy ?? null,
-    templatePath: input.result.executionIntegrity?.templatePath ?? null,
-    taskValidationHint: input.result.verification.taskValidationHint,
-    effectiveValidationCommand: input.result.verification.effectiveValidationCommand,
-    normalizedValidationCommandFrom: input.result.verification.normalizedValidationCommandFrom,
-    executionStatus: input.result.executionStatus,
-    executionMessage: input.result.execution.message ?? null,
-    verificationStatus: input.result.verificationStatus,
-    completionClassification: input.result.completionClassification,
-    backlog: input.result.backlog,
-    followUpAction: input.result.followUpAction,
-    stopReason: input.result.stopReason,
-    remediation: input.result.remediation,
-    summary: input.result.summary,
-    artifactDir: input.result.artifactDir,
-    summaryPath: input.paths.summaryPath,
-    promptPath: input.paths.promptPath,
-    promptEvidencePath: input.paths.promptEvidencePath,
-    executionPlanPath: input.result.executionIntegrity?.executionPlanPath ?? input.paths.executionPlanPath,
-    cliInvocationPath: input.result.executionIntegrity?.cliInvocationPath,
-    promptArtifactPath: input.result.executionIntegrity?.promptArtifactPath ?? input.paths.promptPath,
-    promptHash: input.result.executionIntegrity?.promptHash ?? null,
-    executionPlanHash: input.result.executionIntegrity?.executionPlanHash ?? null,
-    executionPayloadMatched: input.result.executionIntegrity?.executionPayloadMatched ?? null,
-    transcriptPath: input.result.execution.transcriptPath ?? null,
-    lastMessagePath: input.result.execution.lastMessagePath ?? null,
-    executionSummaryPath: input.paths.executionSummaryPath,
-    verifierSummaryPath: input.paths.verifierSummaryPath,
-    iterationResultPath: input.paths.iterationResultPath,
-    remediationPath: input.result.remediation ? input.paths.remediationPath : null,
-    diffSummaryPath: input.diffSummary ? input.paths.diffSummaryPath : null,
-    stdoutPath: input.paths.stdoutPath,
-    stderrPath: input.paths.stderrPath,
-    completionReportStatus: input.result.completionReportStatus ?? null,
-    reconciliationWarnings: input.result.reconciliationWarnings ?? [],
-    warnings: input.result.warnings,
-    errors: input.result.errors
-  };
+export interface RalphGeneratedArtifactRetentionSummary {
+  deletedIterationDirectories: string[];
+  retainedIterationDirectories: string[];
+  protectedRetainedIterationDirectories: string[];
+  deletedPromptFiles: string[];
+  retainedPromptFiles: string[];
+  protectedRetainedPromptFiles: string[];
+  deletedRunArtifactBaseNames: string[];
+  retainedRunArtifactBaseNames: string[];
+  protectedRetainedRunArtifactBaseNames: string[];
+  deletedHandoffFiles?: string[];
+  retainedHandoffFiles?: string[];
 }
 
 export function resolveIterationArtifactPaths(artifactRootDir: string, iteration: number): RalphIterationArtifactPaths {
@@ -1122,576 +274,6 @@ export async function ensureIterationArtifactDirectory(paths: RalphIterationArti
 
 async function ensureProvenanceBundleDirectory(paths: RalphProvenanceBundlePaths): Promise<void> {
   await fs.mkdir(paths.directory, { recursive: true });
-}
-
-async function readTextRecord(target: string): Promise<string | null> {
-  try {
-    return await fs.readFile(target, 'utf8');
-  } catch {
-    return null;
-  }
-}
-
-function looksLikePreflightRecord(record: Record<string, unknown> | null): boolean {
-  return record?.kind === 'preflight'
-    && typeof record.provenanceId === 'string'
-    && typeof record.iteration === 'number'
-    && typeof record.promptKind === 'string'
-    && typeof record.promptTarget === 'string'
-    && typeof record.trustLevel === 'string'
-    && typeof record.ready === 'boolean'
-    && typeof record.summary === 'string'
-    && typeof record.artifactDir === 'string'
-    && typeof record.reportPath === 'string'
-    && typeof record.createdAt === 'string'
-    && Array.isArray(record.diagnostics);
-}
-
-function looksLikeIntegrityFailure(record: Record<string, unknown> | null): boolean {
-  return record?.kind === 'integrityFailure'
-    && typeof record.provenanceId === 'string'
-    && typeof record.iteration === 'number'
-    && typeof record.promptKind === 'string'
-    && typeof record.promptTarget === 'string'
-    && typeof record.trustLevel === 'string'
-    && typeof record.stage === 'string'
-    && record.blocked === true
-    && typeof record.summary === 'string'
-    && typeof record.message === 'string'
-    && typeof record.artifactDir === 'string'
-    && typeof record.createdAt === 'string';
-}
-
-function looksLikeProvenanceBundle(record: Record<string, unknown> | null): boolean {
-  return record?.kind === 'provenanceBundle'
-    && typeof record.provenanceId === 'string'
-    && typeof record.iteration === 'number'
-    && typeof record.promptKind === 'string'
-    && typeof record.promptTarget === 'string'
-    && typeof record.trustLevel === 'string'
-    && typeof record.status === 'string'
-    && typeof record.summary === 'string'
-    && typeof record.artifactDir === 'string'
-    && typeof record.bundleDir === 'string'
-    && typeof record.preflightReportPath === 'string'
-    && typeof record.preflightSummaryPath === 'string'
-    && typeof record.createdAt === 'string'
-    && typeof record.updatedAt === 'string'
-    && typeof record.rootPolicy === 'object'
-    && record.rootPolicy !== null;
-}
-
-function provenanceIdFromRecord(record: Record<string, unknown> | null): string | null {
-  return typeof record?.provenanceId === 'string' && record.provenanceId.trim().length > 0
-    ? record.provenanceId
-    : null;
-}
-
-function parseIterationFromLatestSurface(raw: string | null): number | null {
-  if (!raw) {
-    return null;
-  }
-
-  const headingMatch = /^# Ralph (?:Iteration|Preflight|Provenance Failure) (\d+)\s*$/m.exec(raw);
-  if (headingMatch) {
-    return Number.parseInt(headingMatch[1], 10);
-  }
-
-  const bulletMatch = /^- Iteration:\s+(\d+)\s*$/m.exec(raw);
-  if (bulletMatch) {
-    return Number.parseInt(bulletMatch[1], 10);
-  }
-
-  return null;
-}
-
-function parseArtifactPathsFromLatestSurface(raw: string | null): string[] {
-  if (!raw) {
-    return [];
-  }
-
-  const pathLabels = new Set([
-    'Prompt',
-    'Report',
-    'Preflight report',
-    'Preflight summary',
-    'Iteration artifact dir'
-  ]);
-  const extractedPaths: string[] = [];
-
-  for (const line of raw.split(/\r?\n/u)) {
-    const match = /^- ([^:]+):\s+(.+?)\s*$/.exec(line);
-    if (!match || !pathLabels.has(match[1])) {
-      continue;
-    }
-
-    extractedPaths.push(match[2].trim());
-  }
-
-  return extractedPaths;
-}
-
-function derivedLatestSurfaceReferences(input: {
-  artifactRootDir: string;
-  latestSummaryText: string | null;
-  latestPreflightSummaryText: string | null;
-  latestProvenanceSummaryText: string | null;
-}): string[] {
-  const surfaces = [
-    input.latestSummaryText,
-    input.latestPreflightSummaryText,
-    input.latestProvenanceSummaryText
-  ];
-  const iterations = surfaces
-    .map((surface) => parseIterationFromLatestSurface(surface))
-    .filter((value): value is number => value !== null);
-  const artifactPaths = surfaces.flatMap((surface) => parseArtifactPathsFromLatestSurface(surface));
-
-  return [
-    ...artifactPaths,
-    ...iterations.map((iteration) => derivedIterationDirectoryPath(input.artifactRootDir, iteration))
-  ];
-}
-
-async function resolveProtectedBundleIds(artifactRootDir: string): Promise<Set<string>> {
-  const latestPaths = resolveLatestArtifactPaths(artifactRootDir);
-  const records = await Promise.all([
-    readJsonRecord(latestPaths.latestResultPath),
-    readJsonRecord(latestPaths.latestPreflightReportPath),
-    readJsonRecord(latestPaths.latestPromptEvidencePath),
-    readJsonRecord(latestPaths.latestExecutionPlanPath),
-    readJsonRecord(latestPaths.latestCliInvocationPath),
-    readJsonRecord(latestPaths.latestProvenanceBundlePath),
-    readJsonRecord(latestPaths.latestProvenanceFailurePath)
-  ]);
-
-  return new Set(records
-    .map((record) => provenanceIdFromRecord(record))
-    .filter((value): value is string => Boolean(value)));
-}
-
-export async function repairLatestArtifactSurfaces(artifactRootDir: string): Promise<RalphLatestArtifactRepairSummary> {
-  const latestPaths = resolveLatestArtifactPaths(artifactRootDir);
-  const repairedLatestArtifactPaths: string[] = [];
-  const staleLatestArtifactPaths: string[] = [];
-  const [
-    latestResultExists,
-    latestSummaryExists,
-    latestPreflightReportRecord,
-    latestPreflightSummaryExists,
-    latestProvenanceBundleRecord,
-    latestProvenanceSummaryExists
-  ] = await Promise.all([
-    pathExists(latestPaths.latestResultPath),
-    pathExists(latestPaths.latestSummaryPath),
-    readJsonRecord(latestPaths.latestPreflightReportPath),
-    pathExists(latestPaths.latestPreflightSummaryPath),
-    readJsonRecord(latestPaths.latestProvenanceBundlePath),
-    pathExists(latestPaths.latestProvenanceSummaryPath)
-  ]);
-
-  if (!latestSummaryExists && latestResultExists) {
-    const latestResultRecord = await readJsonRecord(latestPaths.latestResultPath);
-    const repairedSummary = looksLikePreflightRecord(latestResultRecord)
-      ? renderPreflightSummary(latestResultRecord as unknown as RalphPersistedPreflightReport)
-      : looksLikeIntegrityFailure(latestResultRecord)
-        ? renderIntegrityFailureSummary(latestResultRecord as unknown as RalphIntegrityFailure)
-        : latestResultRecord
-          ? renderLatestResultSummary(latestResultRecord)
-          : null;
-
-    if (repairedSummary) {
-      await fs.writeFile(latestPaths.latestSummaryPath, `${repairedSummary.trimEnd()}\n`, 'utf8');
-      repairedLatestArtifactPaths.push(latestPaths.latestSummaryPath);
-    } else {
-      staleLatestArtifactPaths.push(latestPaths.latestSummaryPath);
-    }
-  }
-
-  if (!latestPreflightSummaryExists && looksLikePreflightRecord(latestPreflightReportRecord)) {
-    const repairedSummary = renderPreflightSummary(latestPreflightReportRecord as unknown as RalphPersistedPreflightReport);
-    await fs.writeFile(latestPaths.latestPreflightSummaryPath, `${repairedSummary.trimEnd()}\n`, 'utf8');
-    repairedLatestArtifactPaths.push(latestPaths.latestPreflightSummaryPath);
-  } else if (!latestPreflightSummaryExists && latestPreflightReportRecord) {
-    staleLatestArtifactPaths.push(latestPaths.latestPreflightSummaryPath);
-  }
-
-  if (!latestProvenanceSummaryExists && looksLikeProvenanceBundle(latestProvenanceBundleRecord)) {
-    const repairedSummary = renderProvenanceSummary(latestProvenanceBundleRecord as unknown as RalphProvenanceBundle);
-    await fs.writeFile(latestPaths.latestProvenanceSummaryPath, `${repairedSummary.trimEnd()}\n`, 'utf8');
-    repairedLatestArtifactPaths.push(latestPaths.latestProvenanceSummaryPath);
-  } else if (!latestProvenanceSummaryExists && latestProvenanceBundleRecord) {
-    staleLatestArtifactPaths.push(latestPaths.latestProvenanceSummaryPath);
-  }
-
-  return {
-    repairedLatestArtifactPaths,
-    staleLatestArtifactPaths
-  };
-}
-
-async function resolveProtectedGeneratedArtifacts(input: {
-  artifactRootDir: string;
-  promptDir: string;
-  runDir: string;
-  stateFilePath: string;
-  protectionScope?: RalphGeneratedArtifactProtectionScope;
-}): Promise<RalphProtectedGeneratedArtifacts> {
-  const latestPaths = resolveLatestArtifactPaths(input.artifactRootDir);
-  const [
-    stateRecord,
-    latestResult,
-    latestPreflightReport,
-    latestPromptEvidence,
-    latestExecutionPlan,
-    latestCliInvocation,
-    latestProvenanceBundle,
-    latestProvenanceFailure,
-    latestSummaryText,
-    latestPreflightSummaryText,
-    latestProvenanceSummaryText
-  ] = await Promise.all([
-    readJsonRecord(input.stateFilePath),
-    readJsonRecord(latestPaths.latestResultPath),
-    readJsonRecord(latestPaths.latestPreflightReportPath),
-    readJsonRecord(latestPaths.latestPromptEvidencePath),
-    readJsonRecord(latestPaths.latestExecutionPlanPath),
-    readJsonRecord(latestPaths.latestCliInvocationPath),
-    readJsonRecord(latestPaths.latestProvenanceBundlePath),
-    readJsonRecord(latestPaths.latestProvenanceFailurePath),
-    readTextRecord(latestPaths.latestSummaryPath),
-    readTextRecord(latestPaths.latestPreflightSummaryPath),
-    readTextRecord(latestPaths.latestProvenanceSummaryPath)
-  ]);
-  const protectedArtifacts = createProtectedGeneratedArtifacts();
-  const pathInput = {
-    artifactRootDir: input.artifactRootDir,
-    promptDir: input.promptDir,
-    runDir: input.runDir
-  };
-
-  addProtectedGeneratedArtifactPaths(protectedArtifacts, pathInput, currentStateArtifactReferences({
-    stateRecord,
-    artifactRootDir: input.artifactRootDir,
-    includeHistory: (input.protectionScope ?? 'fullStateAndLatest') === 'fullStateAndLatest'
-  }));
-  addProtectedGeneratedArtifactPaths(protectedArtifacts, pathInput, latestArtifactReferences({
-    latestResult,
-    latestPreflightReport,
-    latestPromptEvidence,
-    latestExecutionPlan,
-    latestCliInvocation,
-    latestProvenanceBundle,
-    latestProvenanceFailure,
-    latestSummaryText,
-    latestPreflightSummaryText,
-    latestProvenanceSummaryText
-  }, pathInput));
-
-  return protectedArtifacts;
-}
-
-export async function cleanupProvenanceBundles(input: {
-  artifactRootDir: string;
-  retentionCount: number;
-}): Promise<RalphProvenanceRetentionSummary> {
-  const inspection = await collectProvenanceBundleRetentionInspection(input);
-  if (input.retentionCount <= 0) {
-    return {
-      deletedBundleIds: [],
-      retainedBundleIds: inspection.retainedBundleIds,
-      protectedBundleIds: inspection.protectedBundleIds
-    };
-  }
-
-  const runsDir = path.join(input.artifactRootDir, 'runs');
-  const retainedIds = new Set(inspection.retainedBundleIds);
-
-  const deletedBundleIds: string[] = [];
-  for (const bundleId of inspection.bundleIds.slice(input.retentionCount)) {
-    if (retainedIds.has(bundleId)) {
-      continue;
-    }
-
-    await fs.rm(path.join(runsDir, bundleId), { recursive: true, force: true });
-    deletedBundleIds.push(bundleId);
-  }
-
-  return {
-    deletedBundleIds,
-    retainedBundleIds: inspection.retainedBundleIds,
-    protectedBundleIds: inspection.protectedBundleIds
-  };
-}
-
-export async function inspectProvenanceBundleRetention(input: {
-  artifactRootDir: string;
-  retentionCount: number;
-}): Promise<RalphProvenanceRetentionSummary> {
-  const inspection = await collectProvenanceBundleRetentionInspection(input);
-  return {
-    deletedBundleIds: [],
-    retainedBundleIds: inspection.retainedBundleIds,
-    protectedBundleIds: inspection.protectedBundleIds
-  };
-}
-
-export async function cleanupGeneratedArtifacts(input: {
-  artifactRootDir: string;
-  promptDir: string;
-  runDir: string;
-  handoffDir?: string;
-  stateFilePath: string;
-  retentionCount: number;
-  protectionScope?: RalphGeneratedArtifactProtectionScope;
-}): Promise<RalphGeneratedArtifactRetentionSummary> {
-  if (input.retentionCount <= 0) {
-    const summary: RalphGeneratedArtifactRetentionSummary = {
-      deletedIterationDirectories: [],
-      retainedIterationDirectories: [],
-      protectedRetainedIterationDirectories: [],
-      deletedPromptFiles: [],
-      retainedPromptFiles: [],
-      protectedRetainedPromptFiles: [],
-      deletedRunArtifactBaseNames: [],
-      retainedRunArtifactBaseNames: [],
-      protectedRetainedRunArtifactBaseNames: []
-    };
-    if (input.handoffDir) {
-      summary.deletedHandoffFiles = [];
-      summary.retainedHandoffFiles = [];
-    }
-    return summary;
-  }
-  const inspection = await collectGeneratedArtifactRetentionInspection(input);
-
-  const retainedIterationDirectories = inspection.iterationDirectoryDecision.retainedNames;
-  const deletedIterationDirectories: string[] = [];
-  for (const entry of inspection.iterationDirectories.slice(input.retentionCount)) {
-    if (retainedIterationDirectories.has(entry.name)) {
-      continue;
-    }
-    await fs.rm(path.join(input.artifactRootDir, entry.name), { recursive: true, force: true });
-    deletedIterationDirectories.push(entry.name);
-  }
-
-  const retainedPromptFiles = inspection.promptFileDecision.retainedNames;
-  const deletedPromptFiles: string[] = [];
-  for (const entry of inspection.promptFiles.slice(input.retentionCount)) {
-    if (retainedPromptFiles.has(entry.name)) {
-      continue;
-    }
-    await fs.rm(path.join(input.promptDir, entry.name), { force: true });
-    deletedPromptFiles.push(entry.name);
-  }
-
-  const retainedRunArtifactBaseNames = inspection.runArtifactDecision.retainedNames;
-  const deletedRunArtifactBaseNames: string[] = [];
-  for (const entry of inspection.runArtifacts.slice(input.retentionCount)) {
-    if (retainedRunArtifactBaseNames.has(entry.baseName)) {
-      continue;
-    }
-    await Promise.all(entry.fileNames.map((fileName) => fs.rm(path.join(input.runDir, fileName), { force: true })));
-    deletedRunArtifactBaseNames.push(entry.baseName);
-  }
-
-  const summary: RalphGeneratedArtifactRetentionSummary = {
-    deletedIterationDirectories,
-    retainedIterationDirectories: inspection.iterationDirectories
-      .filter((entry) => inspection.iterationDirectoryDecision.retainedNames.has(entry.name))
-      .map((entry) => entry.name),
-    protectedRetainedIterationDirectories: inspection.iterationDirectoryDecision.protectedRetainedNames,
-    deletedPromptFiles,
-    retainedPromptFiles: inspection.promptFiles
-      .filter((entry) => inspection.promptFileDecision.retainedNames.has(entry.name))
-      .map((entry) => entry.name),
-    protectedRetainedPromptFiles: inspection.promptFileDecision.protectedRetainedNames,
-    deletedRunArtifactBaseNames,
-    retainedRunArtifactBaseNames: inspection.runArtifacts
-      .filter((entry) => inspection.runArtifactDecision.retainedNames.has(entry.baseName))
-      .map((entry) => entry.baseName),
-    protectedRetainedRunArtifactBaseNames: inspection.runArtifactDecision.protectedRetainedNames
-  };
-
-  if (input.handoffDir) {
-    const retainedHandoffFiles = inspection.handoffFileDecision.retainedNames;
-    const deletedHandoffFiles: string[] = [];
-    for (const entry of inspection.handoffFiles.slice(input.retentionCount)) {
-      if (retainedHandoffFiles.has(entry.name)) {
-        continue;
-      }
-      await fs.rm(path.join(input.handoffDir, entry.name), { force: true });
-      deletedHandoffFiles.push(entry.name);
-    }
-
-    summary.deletedHandoffFiles = deletedHandoffFiles;
-    summary.retainedHandoffFiles = inspection.handoffFiles
-      .filter((entry) => retainedHandoffFiles.has(entry.name))
-      .map((entry) => entry.name);
-  };
-
-  return summary;
-}
-
-export async function inspectGeneratedArtifactRetention(input: {
-  artifactRootDir: string;
-  promptDir: string;
-  runDir: string;
-  handoffDir?: string;
-  stateFilePath: string;
-  retentionCount: number;
-  protectionScope?: RalphGeneratedArtifactProtectionScope;
-}): Promise<RalphGeneratedArtifactRetentionSummary> {
-  const inspection = await collectGeneratedArtifactRetentionInspection(input);
-  const summary: RalphGeneratedArtifactRetentionSummary = {
-    deletedIterationDirectories: [],
-    retainedIterationDirectories: inspection.iterationDirectories
-      .filter((entry) => inspection.iterationDirectoryDecision.retainedNames.has(entry.name))
-      .map((entry) => entry.name),
-    protectedRetainedIterationDirectories: inspection.iterationDirectoryDecision.protectedRetainedNames,
-    deletedPromptFiles: [],
-    retainedPromptFiles: inspection.promptFiles
-      .filter((entry) => inspection.promptFileDecision.retainedNames.has(entry.name))
-      .map((entry) => entry.name),
-    protectedRetainedPromptFiles: inspection.promptFileDecision.protectedRetainedNames,
-    deletedRunArtifactBaseNames: [],
-    retainedRunArtifactBaseNames: inspection.runArtifacts
-      .filter((entry) => inspection.runArtifactDecision.retainedNames.has(entry.baseName))
-      .map((entry) => entry.baseName),
-    protectedRetainedRunArtifactBaseNames: inspection.runArtifactDecision.protectedRetainedNames
-  };
-
-  if (input.handoffDir) {
-    summary.deletedHandoffFiles = [];
-    summary.retainedHandoffFiles = inspection.handoffFiles
-      .filter((entry) => inspection.handoffFileDecision.retainedNames.has(entry.name))
-      .map((entry) => entry.name);
-  }
-
-  return summary;
-}
-
-async function collectProvenanceBundleRetentionInspection(input: {
-  artifactRootDir: string;
-  retentionCount: number;
-}): Promise<RalphBundleRetentionInspection> {
-  const runsDir = path.join(input.artifactRootDir, 'runs');
-  const entries = await fs.readdir(runsDir, { withFileTypes: true }).catch(() => []);
-  const bundleIds = entries
-    .filter((entry) => entry.isDirectory() && entry.name.startsWith('run-'))
-    .map((entry) => entry.name)
-    .sort((left, right) => right.localeCompare(left));
-  const protectedIds = Array.from(await resolveProtectedBundleIds(input.artifactRootDir)).sort();
-  const retainedIds = new Set(input.retentionCount <= 0 ? bundleIds : bundleIds.slice(0, input.retentionCount));
-  protectedIds.forEach((bundleId) => retainedIds.add(bundleId));
-
-  return {
-    bundleIds,
-    retainedBundleIds: bundleIds.filter((bundleId) => retainedIds.has(bundleId)),
-    protectedBundleIds: protectedIds
-  };
-}
-
-async function collectGeneratedArtifactRetentionInspection(input: {
-  artifactRootDir: string;
-  promptDir: string;
-  runDir: string;
-  handoffDir?: string;
-  stateFilePath: string;
-  retentionCount: number;
-  protectionScope?: RalphGeneratedArtifactProtectionScope;
-}): Promise<RalphGeneratedArtifactRetentionInspection> {
-  const [artifactEntries, promptEntries, runEntries, protectedArtifacts] = await Promise.all([
-    fs.readdir(input.artifactRootDir, { withFileTypes: true }).catch(() => []),
-    fs.readdir(input.promptDir, { withFileTypes: true }).catch(() => []),
-    fs.readdir(input.runDir, { withFileTypes: true }).catch(() => []),
-    resolveProtectedGeneratedArtifacts({
-      artifactRootDir: input.artifactRootDir,
-      promptDir: input.promptDir,
-      runDir: input.runDir,
-      stateFilePath: input.stateFilePath,
-      protectionScope: input.protectionScope
-    })
-  ]);
-  const handoffEntries = input.handoffDir
-    ? await fs.readdir(input.handoffDir, { withFileTypes: true }).catch(() => [])
-    : [];
-
-  const iterationDirectories = artifactEntries
-    .filter((entry) => entry.isDirectory())
-    .map((entry) => parseIterationDirectoryName(entry.name))
-    .filter((entry): entry is { iteration: number; name: string } => entry !== null)
-    .sort(sortByIterationDesc);
-  const promptFiles = promptEntries
-    .filter((entry) => entry.isFile())
-    .map((entry) => parsePromptFileName(entry.name))
-    .filter((entry): entry is { iteration: number; name: string } => entry !== null)
-    .sort(sortByIterationDesc);
-  const runArtifactGroups = new Map<string, { baseName: string; iteration: number; fileNames: string[] }>();
-  for (const entry of runEntries.filter((candidate) => candidate.isFile())) {
-    const parsed = parseRunArtifactFileName(entry.name);
-    if (!parsed) {
-      continue;
-    }
-
-    const current = runArtifactGroups.get(parsed.baseName);
-    if (current) {
-      current.fileNames.push(parsed.name);
-      continue;
-    }
-
-    runArtifactGroups.set(parsed.baseName, {
-      baseName: parsed.baseName,
-      iteration: parsed.iteration,
-      fileNames: [parsed.name]
-    });
-  }
-
-  const runArtifacts = Array.from(runArtifactGroups.values()).sort((left, right) =>
-    right.iteration - left.iteration || right.baseName.localeCompare(left.baseName)
-  );
-  const handoffFiles = handoffEntries
-    .filter((entry) => entry.isFile())
-    .map((entry) => parseHandoffFileName(entry.name))
-    .filter((entry): entry is { iteration: number; name: string } => entry !== null)
-    .sort(sortByIterationDesc);
-  const effectiveRetentionCount = input.retentionCount <= 0
-    ? Math.max(iterationDirectories.length, promptFiles.length, runArtifacts.length, handoffFiles.length)
-    : input.retentionCount;
-
-  return {
-    iterationDirectories,
-    promptFiles,
-    runArtifacts,
-    handoffFiles,
-    protectedArtifacts,
-    iterationDirectoryDecision: retentionDecisionByNewestAndProtected(
-      iterationDirectories,
-      effectiveRetentionCount,
-      protectedArtifacts.iterationDirectories,
-      (entry) => entry.name
-    ),
-    promptFileDecision: retentionDecisionByNewestAndProtected(
-      promptFiles,
-      effectiveRetentionCount,
-      protectedArtifacts.promptFiles,
-      (entry) => entry.name
-    ),
-    runArtifactDecision: retentionDecisionByNewestAndProtected(
-      runArtifacts,
-      effectiveRetentionCount,
-      protectedArtifacts.runArtifactBaseNames,
-      (entry) => entry.baseName
-    ),
-    handoffFileDecision: {
-      retainedNames: retainedNamesByNewestAndProtected(
-        handoffFiles,
-        effectiveRetentionCount,
-        [],
-        (entry) => entry.name
-      )
-    }
-  };
 }
 
 export async function writePromptArtifacts(input: {
